+++ source/changelog.md
@@ -1,6 +1,9 @@
 ﻿ChanSort Change Log 
 ===================
 
+2023-01-10
+- TCL: fixed deleting channels
+
 2023-01-08
 - TCL/Thomson .tar: custom implementation for reading/writing .tar archives, preserving all 
   unix file metadata (based on "old-GNU" .tar flavor, like the files exported by the TV)